requirements.txt
streamlit
google-cloud-vision
google-auth
